Moved speakers to assets folder
The /public folder should not contain assets like images - these should be in the assets folder. I've moved the speaker photos there, and replaced spaces in the file name with underscores (spaces in file names is best avoided, and can lead to problems in different browsers).

To reference the image in the JSON config use the path from the asset directory. In the case for speakers it would be "img/speakers/cropped/speaker_name.jpg". 

This is so we avoid hard linking to images — i.e. when the site domain changes, each url will not break.